+++ packages/iocuak/src/container/models/api/BindValueOptionsApi.ts
@@ -1,0 +1,8 @@
+import { BindingTag } from '../../../binding/models/domain/BindingTag';
+import { ServiceId } from '../../../common/models/domain/ServiceId';
+
+export interface BindValueOptionsApi {
+  serviceId: ServiceId;
+  tags?: BindingTag | BindingTag[];
+  value: unknown;
+}