+++ packages/iocuak/src/container/utils/bindToValue.ts
@@ -1,3 +1,4 @@
+import { BindingTag } from '../../binding/models/domain/BindingTag';
 import { BindingType } from '../../binding/models/domain/BindingType';
 import { ValueBinding } from '../../binding/models/domain/ValueBinding';
 import { BindingService } from '../../binding/services/domain/BindingService';
@@ -5,13 +6,14 @@ import { ServiceId } from '../../common/models/domain/ServiceId';
 
 export function bindToValue<TInstance>(
   serviceId: ServiceId,
+  tags: BindingTag[],
   value: TInstance,
   containerBindingService: BindingService,
 ): void {
   const valueBinding: ValueBinding<TInstance> = {
     bindingType: BindingType.value,
     id: serviceId,
-    tags: [],
+    tags: [...tags],
     value: value,
   };